Use existing filename compression function if name too long (#600)
Previous filename compression technique was to always trim the buildname to
only the first 80 characters as to avoid "filename too long" errors.

Cache file or directory names are built in the format of `testName_buildName`
The above method does not protect against the case where testName may be
very long.

This implementation uses an existing function named `getCompressedFileNameIfTooLong` in
`CDashQueryAnalyzeReport.py` module file which will form a hash of the passed in string
if it is deamed too long.

This will also help mitigate the chances of a filename collision as previously it was
possible for a trimmed buildName to result in the same `testName_buildName` filename
if testName was the same test and had the correct length.