Bring back Copyright.txt file since it is assumed/required by TriBITS builds/tests
This is a fix from the last git format-patch/am of a commit created on the
Trilinos side.

The Copyright.txt file is expected in some of the TriBITS example project
builds.

The file:

* tribits/Copyright.txt

 is created as a symlink from tribits/COPYRIGHT added in the laste commit.

The files:

* tribits/examples/MockTrilinos/Copyright.txt
* tribits/examples/TribitsExampleProject/Copyright.txt

were deleted in the last git format-path/am commit and are added back here
from what is on TriBITS 'master'.