Add new function tribits_package_is_not_explicitly_disabled()
This makes the logic in the code a little more understandable I think.  I was
looking over the code and I thought that this could be improved.  This will
get used in more code in future commits.  NOTE: This gets tested through
existing tests by usage in higher-level code (like the DependencyUnitTests
tests).

We will also need other functions like this to make the code more
self-documenting.

I also:

* Added a tailing impl comment to help try to explain something.

* Put the documented functions back in alphabetical order (moved
  tribits_process_enabled_standard_tpl()).