Adjust clone tests for old Python implementation (#610)
These tests clone older versions of TribitsExampleProject that are still using
logic to disable the WrapExternal package based on PYTHON_EXECUTABLE.
Therefore, with new TriBITS implementation setting Python3_EXECUTABLE, the old
logic does not allow the enable of WrapExternal and therefore there is one
less test (10 vs. 11).

This commit adjusts the regex for the number of tests being run for that
change.

Once these external repo snapshots are updated, then we will need to revert
this commit (and fix the SHA1s as usual).

Signed-off-by: Roscoe A. Bartlett <rabartl@sandia.gov>