Reduce min Python version from 3.8 to 3.6 (#610)
The default version of `python3` on RHEL8 machines in Python 3.6.8.  (I am not
sure where I got Python 3.8 from.)

But note that RHEL8 machines also seem to have Python 3.11 and 3.12 installed
as well.  I am wondering if it would not be more wise to be pro-active and
push the version of Python being required to a higher version that will be
supported for longer.  (Python 3.6 ended support at the end of 2021.)